chore: useAuth hook 폴더 이하로 이동
