+++ client/src/shared/context/AuthContext.tsx
@@ -1,11 +1,11 @@
-import React, { createContext, useContext } from "react";
+import React, { createContext } from "react";
 import type { UserType } from "@/shared/types/auth";
 
-interface AuthContextType {
+export interface AuthContextType {
   userType: UserType;
 }
 
-const AuthContext = createContext<AuthContextType | null>(null);
+export const AuthContext = createContext<AuthContextType | null>(null);
 
 interface AuthProviderProps {
   children: React.ReactNode;
@@ -19,13 +19,4 @@ export const AuthProvider: React.FC<AuthProviderProps> = ({
   return (
     <AuthContext.Provider value={{ userType }}>{children}</AuthContext.Provider>
   );
-};
-
-// Hook 생성
-export const useAuth = (): AuthContextType => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error();
-  }
-  return context;
 };
