+++ client/src/shared/hook/useAuth.tsx
@@ -1,0 +1,13 @@
+import { useContext } from "react";
+import {
+  AuthContext,
+  type AuthContextType,
+} from "@/shared/context/AuthContext";
+
+export const useAuth = (): AuthContextType => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error();
+  }
+  return context;
+};
